+++ R/qol.R
@@ -62,8 +62,8 @@
 #'
 #' Reporting: [content_report()]
 #'
-#' Small helpers: [setcolorder_by_pattern()], [drop_type_vars()],
-#' [fuse_variables()], [get_excel_range()], [get_integer_length()]
+#' Small helpers: [setcolorder_by_pattern()], [drop_type_vars()], [fuse_variables()],
+#' [get_excel_range()], [get_integer_length()], [round_values()]
 #'
 #' Split data frame: [split_by]
 #'